Add files via upload
The first visual is scatter plot. The scatter plot is to portray the correlation of every player's weight and the number of touchdowns. The purpose is to display that there really isn't much correlation.

The second visual is a bar graph. I would categorize the weight into groups. For instance, one data would be from 200-204 pounds. The next is 205-209 pounds and so on. The purpose is to show how many touchdowns the players scored based on their weight groups.